feat(vm): extract triggeractionviewmodel
